refactor(core): Split the workflow review request service (no-changelog)
Move the reviewer-facing read side of workflow reviews — the cross-project
inbox, its counts, and the review detail — into WorkflowReviewInboxService.
Its visibility starts from the user's projects, while what remains in
WorkflowReviewRequestService starts from a known workflow, so the two halves
had almost disjoint dependencies.

Replace the two competing feature-gate idioms in the module with a single
WorkflowReviewFeatureGate: the five lifecycle methods each inlined only the
policy check, while the inbox methods checked license and policy. All entry
points now check both, which also gates any future non-HTTP caller.

No route, API-type, or behavior change. Every moved and retained method is
byte-identical apart from the gate call, and all unit and integration tests
pass unchanged.
