+++ packages/cli/src/modules/workflow-reviews.ee/workflow-review-request.service.ts
@@ -7,64 +7,54 @@ import type {
 	WorkflowReviewEligibleReviewersList,
 	WorkflowReviewRequestList,
 	WorkflowReviewRequestSummary,
-	ListWorkflowReviewInboxQueryDto,
-	ListWorkflowReviewInboxResponse,
-	GetWorkflowReviewInboxSummaryResponse,
-	WorkflowReviewInboxItem,
 	WorkflowReviewEligibleReviewer,
-	WorkflowReviewRequestDetail,
-	WorkflowReviewRequestWorkflowDetail,
-	WorkflowReviewVersionSnapshot,
 } from '@n8n/api-types';
-import { LicenseState, Logger } from '@n8n/backend-common';
+import { Logger } from '@n8n/backend-common';
 import {
 	DbLock,
 	DbLockService,
 	ProjectRelationRepository,
 	SharedWorkflowRepository,
 	UserRepository,
-	WorkflowPublishedVersionRepository,
 	WorkflowReviewRequestAuthorRepository,
 	WorkflowReviewRequestRepository,
 	WorkflowReviewRequestReviewerRepository,
 	WorkflowReviewRequestWorkflowRepository,
-	type InboxCursor,
 	type User,
-	type WorkflowHistory,
 	type WorkflowReviewRequest,
-	type WorkflowReviewRequestLinkedWorkflow,
-	type WorkflowReviewRequestReviewer,
-	type WorkflowReviewRequestWorkflowDetailRow,
 } from '@n8n/db';
 import { Service } from '@n8n/di';
 import {
 	GLOBAL_ADMIN_ROLE_SLUG,
 	GLOBAL_OWNER_ROLE_SLUG,
 	PROJECT_ADMIN_ROLE_SLUG,
-	hasGlobalScope,
 } from '@n8n/permissions';
 
 import { CollaborationService } from '@/collaboration/collaboration.service';
-import { isWorkflowReviewsFeatureAvailable } from '@/constants/workflow-reviews';
 import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 import { ConflictError } from '@/errors/response-errors/conflict.error';
 import { ForbiddenError } from '@/errors/response-errors/forbidden.error';
 import { NotFoundError } from '@/errors/response-errors/not-found.error';
-import { ProjectService } from '@/services/project.service.ee';
 import { RoleService } from '@/services/role.service';
-import { WorkflowReviewPolicyService } from '@/services/workflow-review-policy.service';
 import { WorkflowFinderService } from '@/workflows/workflow-finder.service';
 import { WorkflowHistoryService } from '@/workflows/workflow-history/workflow-history.service';
 
+import { WorkflowReviewFeatureGate } from './workflow-review-feature-gate.service';
+
+/**
+ * The workflow-scoped review request lifecycle: listing a workflow's reviews,
+ * resolving its eligible reviewers, opening a review, re-pinning it to a new
+ * version, and deciding it. The reviewer-facing read side lives in
+ * `WorkflowReviewInboxService`.
+ */
 @Service()
 export class WorkflowReviewRequestService {
 	constructor(
 		private readonly logger: Logger,
-		private readonly workflowReviewPolicyService: WorkflowReviewPolicyService,
+		private readonly featureGate: WorkflowReviewFeatureGate,
 		private readonly workflowFinderService: WorkflowFinderService,
 		private readonly workflowHistoryService: WorkflowHistoryService,
 		private readonly sharedWorkflowRepository: SharedWorkflowRepository,
-		private readonly workflowPublishedVersionRepository: WorkflowPublishedVersionRepository,
 		private readonly workflowReviewRequestRepository: WorkflowReviewRequestRepository,
 		private readonly workflowReviewRequestWorkflowRepository: WorkflowReviewRequestWorkflowRepository,
 		private readonly workflowReviewRequestAuthorRepository: WorkflowReviewRequestAuthorRepository,
@@ -72,8 +62,6 @@ export class WorkflowReviewRequestService {
 		private readonly userRepository: UserRepository,
 		private readonly projectRelationRepository: ProjectRelationRepository,
 		private readonly roleService: RoleService,
-		private readonly projectService: ProjectService,
-		private readonly licenseState: LicenseState,
 		private readonly dbLockService: DbLockService,
 		private readonly collaborationService: CollaborationService,
 	) {}
@@ -99,10 +87,7 @@ export class WorkflowReviewRequestService {
 		user: User,
 		query: ListWorkflowReviewRequestsQueryDto,
 	): Promise<WorkflowReviewRequestList> {
-		const policy = await this.workflowReviewPolicyService.get();
-		if (!policy.enabled) {
-			throw new ForbiddenError('Workflow reviews are not enabled for this instance');
-		}
+		await this.featureGate.assertAvailable();
 
 		const workflow = await this.workflowFinderService.findWorkflowForUser(query.workflowId, user, [
 			'workflow:read',
@@ -133,10 +118,7 @@ export class WorkflowReviewRequestService {
 		user: User,
 		query: GetWorkflowReviewEligibleReviewersQueryDto,
 	): Promise<WorkflowReviewEligibleReviewersList> {
-		const policy = await this.workflowReviewPolicyService.get();
-		if (!policy.enabled) {
-			throw new ForbiddenError('Workflow reviews are not enabled for this instance');
-		}
+		await this.featureGate.assertAvailable();
 
 		const workflow = await this.workflowFinderService.findWorkflowForUser(query.workflowId, user, [
 			'workflow:publish',
@@ -175,10 +157,7 @@ export class WorkflowReviewRequestService {
 	): Promise<WorkflowReviewRequestSummary> {
 		const { workflowId, workflowVersionId } = dto.workflows[0];
 
-		const policy = await this.workflowReviewPolicyService.get();
-		if (!policy.enabled) {
-			throw new ForbiddenError('Workflow reviews are not enabled for this instance');
-		}
+		await this.featureGate.assertAvailable();
 
 		const workflow = await this.workflowFinderService.findWorkflowForUser(workflowId, user, [
 			'workflow:publish',
@@ -287,10 +266,7 @@ export class WorkflowReviewRequestService {
 		workflowReviewRequestId: string,
 		dto: UpdateWorkflowReviewRequestVersionDto,
 	): Promise<WorkflowReviewRequestSummary> {
-		const policy = await this.workflowReviewPolicyService.get();
-		if (!policy.enabled) {
-			throw new ForbiddenError('Workflow reviews are not enabled for this instance');
-		}
+		await this.featureGate.assertAvailable();
 
 		const request = await this.workflowReviewRequestRepository.findById(workflowReviewRequestId);
 		if (!request) {
@@ -403,10 +379,7 @@ export class WorkflowReviewRequestService {
 		workflowReviewRequestId: string,
 		dto: DecideWorkflowReviewRequestDto,
 	): Promise<WorkflowReviewRequestSummary> {
-		const policy = await this.workflowReviewPolicyService.get();
-		if (!policy.enabled) {
-			throw new ForbiddenError('Workflow reviews are not enabled for this instance');
-		}
+		await this.featureGate.assertAvailable();
 
 		const request = await this.workflowReviewRequestRepository.findById(workflowReviewRequestId);
 		if (!request) {
@@ -560,320 +533,6 @@ export class WorkflowReviewRequestService {
 			workflowVersionId,
 			createdAt: request.createdAt.toISOString(),
 			updatedAt: request.updatedAt.toISOString(),
-		};
-	}
-
-	/**
-	 * Cross-project inbox.
-	 *
-	 * Defaults to open requests when `state` is omitted. Deferred (LIGO-594):
-	 * `projectId`, `reviewer`, and `author` filters.
-	 */
-	async listForInbox(
-		user: User,
-		query: ListWorkflowReviewInboxQueryDto,
-	): Promise<ListWorkflowReviewInboxResponse> {
-		await this.assertFeatureAvailable();
-
-		const projectIds = await this.resolveAccessibleProjectIds(user);
-		const { limit } = query;
-		const rows = await this.workflowReviewRequestRepository.findManyForInbox({
-			projectIds,
-			requesterId: user.id,
-			state: query.state ?? 'open',
-			limit: limit + 1,
-			cursor: query.cursor ? this.decodeInboxCursor(query.cursor) : undefined,
-		});
-
-		const hasMore = rows.length > limit;
-		const data = rows.slice(0, limit);
-		const lastRow = data.at(-1);
-		const nextCursor = hasMore && lastRow ? this.encodeInboxCursor(lastRow) : null;
-		const requestIds = data.map((row) => row.id);
-		const [linkedWorkflowByRequestId, reviewerRows] = await Promise.all([
-			this.workflowReviewRequestWorkflowRepository.findLinkedWorkflowsByRequestIds(requestIds),
-			this.workflowReviewRequestReviewerRepository.findByRequestIds(requestIds),
-		]);
-
-		const participantsByRequestId = await this.hydrateParticipants(data, reviewerRows);
-
-		return {
-			data: data.map((row) => {
-				const { requester, reviewers } = participantsByRequestId.get(row.id) ?? {
-					requester: null,
-					reviewers: [],
-				};
-				return this.toInboxItem(
-					row,
-					linkedWorkflowByRequestId.get(row.id) ?? null,
-					requester,
-					reviewers,
-				);
-			}),
-			nextCursor,
-			hasMore,
-		};
-	}
-
-	/**
-	 * Batch-resolve the requester and requested reviewers for each request row,
-	 * keyed by request id. Deleted users simply drop out of the result.
-	 */
-	private async hydrateParticipants(
-		rows: WorkflowReviewRequest[],
-		reviewerRows: WorkflowReviewRequestReviewer[],
-	): Promise<
-		Map<
-			string,
-			{
-				requester: WorkflowReviewEligibleReviewer | null;
-				reviewers: WorkflowReviewEligibleReviewer[];
-			}
-		>
-	> {
-		const reviewerIdsByRequestId = new Map<string, string[]>();
-		for (const { workflowReviewRequestId, userId } of reviewerRows) {
-			const ids = reviewerIdsByRequestId.get(workflowReviewRequestId) ?? [];
-			ids.push(userId);
-			reviewerIdsByRequestId.set(workflowReviewRequestId, ids);
-		}
-
-		const userIds = new Set([
-			...rows.map((row) => row.createdById).filter((id) => id !== null),
-			...reviewerRows.map((row) => row.userId),
-		]);
-
-		const usersById = new Map<string, WorkflowReviewEligibleReviewer>();
-		if (userIds.size > 0) {
-			for (const user of await this.userRepository.findManyByIds([...userIds])) {
-				usersById.set(user.id, this.toEligibleReviewer(user));
-			}
-		}
-
-		return new Map(
-			rows.map((row) => [
-				row.id,
-				{
-					requester: row.createdById ? (usersById.get(row.createdById) ?? null) : null,
-					reviewers: (reviewerIdsByRequestId.get(row.id) ?? [])
-						.map((userId) => usersById.get(userId))
-						.filter((reviewer) => reviewer !== undefined),
-				},
-			]),
-		);
-	}
-
-	async getInboxSummaryForUser(user: User): Promise<GetWorkflowReviewInboxSummaryResponse> {
-		await this.assertFeatureAvailable();
-
-		const projectIds = await this.resolveAccessibleProjectIds(user);
-		return await this.workflowReviewRequestRepository.countByStateForInbox({
-			projectIds,
-			requesterId: user.id,
-		});
-	}
-
-	/**
-	 * Visibility starts from the inbox rule (requester OR `workflow:publish` in the
-	 * review's project OR globally), then narrows per workflow to what the caller can
-	 * currently read — see {@link filterReadableWorkflowRows}.
-	 */
-	async getDetail(
-		user: User,
-		workflowReviewRequestId: string,
-	): Promise<WorkflowReviewRequestDetail> {
-		await this.assertFeatureAvailable();
-
-		const request = await this.workflowReviewRequestRepository.findById(workflowReviewRequestId);
-		if (!request || !(await this.canAccessRequest(user, request))) {
-			throw new NotFoundError('Could not find review request');
-		}
-
-		const [workflowRows, reviewerRows] = await Promise.all([
-			this.workflowReviewRequestWorkflowRepository.findLinkedWorkflowDetailsByRequestId(request.id),
-			this.workflowReviewRequestReviewerRepository.findByRequestIds([request.id]),
-		]);
-
-		const readableRows = await this.filterReadableWorkflowRows(user, workflowRows);
-		// Someone who reaches this review through its project has no reason to learn it
-		// exists once they can read none of the workflows it covers. The requester already
-		// knows, and their inbox still lists it, so they keep the record — narrowed to the
-		// workflows they can currently read.
-		if (request.createdById !== user.id && workflowRows.length > 0 && readableRows.length === 0) {
-			throw new NotFoundError('Could not find review request');
-		}
-
-		const [workflows, participantsByRequestId] = await Promise.all([
-			Promise.all(readableRows.map(async (row) => await this.toWorkflowDetail(row))),
-			this.hydrateParticipants([request], reviewerRows),
-		]);
-
-		const { requester, reviewers } = participantsByRequestId.get(request.id) ?? {
-			requester: null,
-			reviewers: [],
-		};
-		return {
-			// One workflow per review for now, so the summary fields mirror the first row
-			...this.toInboxItem(request, workflows.at(0) ?? null, requester, reviewers),
-			description: request.description,
-			workflows,
-		};
-	}
-
-	/** Inbox visibility rule: requester, or `workflow:publish` in the review's project. */
-	private async canAccessRequest(user: User, request: WorkflowReviewRequest): Promise<boolean> {
-		if (request.createdById === user.id) {
-			return true;
-		}
-
-		const projectIds = await this.resolveAccessibleProjectIds(user);
-		return projectIds === null || projectIds.includes(request.projectId);
-	}
-
-	/**
-	 * A review's `projectId` is fixed at creation and nothing closes open reviews when a
-	 * workflow is transferred, so the stored project does not prove the caller may still
-	 * read a covered workflow. Re-check every row against the workflow's *current* owner
-	 * before returning its content.
-	 *
-	 * This applies to the requester too. They held publish rights when they opened the
-	 * review, but may have lost them since — and because the baseline is resolved at read
-	 * time, an exemption would leave them reading versions published after they lost
-	 * access.
-	 */
-	private async filterReadableWorkflowRows(
-		user: User,
-		rows: WorkflowReviewRequestWorkflowDetailRow[],
-	): Promise<WorkflowReviewRequestWorkflowDetailRow[]> {
-		const readable = await Promise.all(
-			rows.map(async (row) =>
-				(await this.workflowFinderService.findWorkflowForUser(row.workflowId, user, [
-					'workflow:read',
-				]))
-					? row
-					: null,
-			),
-		);
-
-		return readable.filter((row): row is WorkflowReviewRequestWorkflowDetailRow => row !== null);
-	}
-
-	/**
-	 * Both diff sides for one child row. The baseline is resolved at read time, so
-	 * a publish during an open review moves what reviewers are diffing against.
-	 */
-	private async toWorkflowDetail(
-		row: WorkflowReviewRequestWorkflowDetailRow,
-	): Promise<WorkflowReviewRequestWorkflowDetail> {
-		const publishedVersionId = await this.workflowPublishedVersionRepository.getPublishedVersionId(
-			row.workflowId,
-		);
-
-		const [pinnedVersion, baselineVersion] = await Promise.all([
-			this.findVersionSnapshot(row.workflowId, row.workflowVersionId),
-			this.findVersionSnapshot(row.workflowId, publishedVersionId),
-		]);
-
-		return {
-			workflowId: row.workflowId,
-			workflowName: row.workflowName,
-			workflowVersionId: row.workflowVersionId,
-			pinnedVersion,
-			baselineVersion,
-		};
-	}
-
-	/** `null` version id, or a version whose history row was pruned, both mean "no content". */
-	private async findVersionSnapshot(
-		workflowId: string,
-		versionId: string | null,
-	): Promise<WorkflowReviewVersionSnapshot | null> {
-		if (!versionId) {
-			return null;
-		}
-
-		const version = await this.workflowHistoryService.findVersion(workflowId, versionId);
-		return version ? this.toVersionSnapshot(version) : null;
-	}
-
-	private toVersionSnapshot(version: WorkflowHistory): WorkflowReviewVersionSnapshot {
-		return {
-			versionId: version.versionId,
-			nodes: version.nodes,
-			connections: version.connections,
-			nodeGroups: version.nodeGroups,
-			createdAt: version.createdAt.toISOString(),
-		};
-	}
-
-	private async assertFeatureAvailable(): Promise<void> {
-		if (!isWorkflowReviewsFeatureAvailable(this.licenseState.isWorkflowReviewsLicensed())) {
-			throw new ForbiddenError('Workflow reviews are not available on this instance');
-		}
-
-		const policy = await this.workflowReviewPolicyService.get();
-		if (!policy.enabled) {
-			throw new ForbiddenError('Workflow reviews are disabled on this instance');
-		}
-	}
-
-	/**
-	 * Project IDs for inbox queries. `null` means "all projects, unfiltered" —
-	 * correct for users with `workflow:publish` scoped globally. Requesters always
-	 * see their own reviews regardless (repository OR-matches `requesterId`), so no
-	 * personal-project fallback is needed.
-	 */
-	async resolveAccessibleProjectIds(user: User): Promise<string[] | null> {
-		if (hasGlobalScope(user, 'workflow:publish')) {
-			return null;
-		}
-
-		return await this.projectService.getProjectIdsWithScope(user, ['workflow:publish']);
-	}
-
-	/**
-	 * Encode the keyset boundary (createdAt + id) into an opaque cursor so the
-	 * next page is resolved without re-reading the anchor row — a review deleted
-	 * between requests no longer truncates the rest of the inbox.
-	 */
-	private encodeInboxCursor(row: WorkflowReviewRequest): string {
-		return Buffer.from(`${row.createdAt.toISOString()}|${row.id}`, 'utf8').toString('base64url');
-	}
-
-	private decodeInboxCursor(cursor: string): InboxCursor {
-		const decoded = Buffer.from(cursor, 'base64url').toString('utf8');
-		const separatorIndex = decoded.indexOf('|');
-		if (separatorIndex === -1) {
-			throw new BadRequestError('Invalid pagination cursor');
-		}
-
-		const createdAt = new Date(decoded.slice(0, separatorIndex));
-		const id = decoded.slice(separatorIndex + 1);
-		if (id.length === 0 || Number.isNaN(createdAt.getTime())) {
-			throw new BadRequestError('Invalid pagination cursor');
-		}
-
-		return { createdAt, id };
-	}
-
-	private toInboxItem(
-		entity: WorkflowReviewRequest,
-		linkedWorkflow: WorkflowReviewRequestLinkedWorkflow | null,
-		requester: WorkflowReviewEligibleReviewer | null,
-		reviewers: WorkflowReviewEligibleReviewer[],
-	): WorkflowReviewInboxItem {
-		return {
-			id: entity.id,
-			projectId: entity.projectId,
-			title: entity.title,
-			workflowName: linkedWorkflow?.workflowName ?? null,
-			workflowVersionId: linkedWorkflow?.workflowVersionId ?? null,
-			decision: entity.decision,
-			state: entity.state,
-			createdAt: entity.createdAt.toISOString(),
-			updatedAt: entity.updatedAt.toISOString(),
-			requester,
-			reviewers,
 		};
 	}
 }
